// src/types/products.tsx
export interface Iproduct {
    id: number;
    _id: number;
    name: string;
    price: number;
    author: string;
    des: string;
    categoryId: number | string;
}

// export interface IRegister {
//     name: string;
//     email: string;
//     password: string;
//     confirmPassword: string;
// }

// export interface ILogin {
//     email: string;
//     password: string;
// }